Remove adding noise on fake objects
Images processed using the LSST software stack are expected to be
background noise dominated for faint objects. As such their is no
reason to add additional noise to fake objects inserted. This also
takes care of a problem where the noise being inserted was
artificially too large due to a bug where the gain was not taken
into account.